refactor label prop socket events
